requirements.txt
langchain==0.1.1
langchain_groq==0.1.1
langchain_community==0.1.1
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.0
gradio==4.12.0
huggingface_hub==0.19.4
sentence-transformers==2.2.2
